requirements.txt
opencv-python-headless>=4.5
paddlepaddle>=2.4.2   # or paddlepaddle-cpu if GPU not available
paddleocr>=2.5
easyocr>=1.6
pytesseract>=0.3.10
numpy>=1.21
streamlit>=1.20
Pillow>=9.0
matplotlib